# docs/swagger.yaml
basePath: /
definitions:
  dto.AuthorResponse:
    properties:
      bio:
        example: British author, best known for Harry Potter
        type: string
      birth_date:
        example: "1965-07-31"
        type: string
      created_at:
        example: "2024-01-15T10:00:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: J.K. Rowling
        type: string
      updated_at:
        example: "2024-01-15T10:00:00Z"
        type: string
    type: object
  dto.BaseResponse:
    properties:
      error:
        example: false
        type: boolean
      msg:
        example: Success
        type: string
    type: object
  dto.BookResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      isbn:
        type: string
      publish_year:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  dto.BookWithAuthorResponse:
    properties:
      author:
        $ref: '#/definitions/dto.AuthorResponse'
      author_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      isbn:
        type: string
      publish_year:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  dto.CreateAuthorRequest:
    properties:
      bio:
        maxLength: 500
        minLength: 10
        type: string
      birth_date:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - bio
    - birth_date
    - name
    type: object
  dto.DataResponse-models_Author:
    properties:
      data:
        $ref: '#/definitions/models.Author'
      error:
        example: false
        type: boolean
      msg:
        example: Success
        type: string
    type: object
  dto.GetAuthorsResponse:
    properties:
      count:
        example: 5
        type: integer
      data:
        items:
          $ref: '#/definitions/dto.AuthorResponse'
        type: array
      error:
        example: false
        type: boolean
      msg:
        example: Success
        type: string
    type: object
  dto.GetBooksResponse:
    properties:
      count:
        example: 5
        type: integer
      data:
        items:
          $ref: '#/definitions/dto.BookResponse'
        type: array
      error:
        example: false
        type: boolean
      msg:
        example: Success
        type: string
    type: object
  dto.GetBooksWithAuthorResponse:
    properties:
      count:
        example: 5
        type: integer
      data:
        items:
          $ref: '#/definitions/dto.BookWithAuthorResponse'
        type: array
      error:
        example: false
        type: boolean
      msg:
        example: Success
        type: string
    type: object
  dto.GetUsersResponse:
    properties:
      count:
        example: 5
        type: integer
      data:
        items:
          $ref: '#/definitions/dto.UserResponse'
        type: array
      error:
        example: false
        type: boolean
      msg:
        example: Success
        type: string
    type: object
  dto.LoginAuthDto:
    properties:
      password:
        maxLength: 100
        minLength: 6
        type: string
      username:
        maxLength: 100
        minLength: 6
        type: string
    required:
    - password
    - username
    type: object
  dto.LoginResponse:
    properties:
      token:
        type: string
    type: object
  dto.UpdateAuthorRequest:
    properties:
      bio:
        maxLength: 500
        minLength: 10
        type: string
      birth_date:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
    type: object
  dto.UserResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_admin:
        type: boolean
      token_version:
        type: integer
      username:
        type: string
    type: object
  dto.ValidationErrorResponse:
    properties:
      error:
        example: true
        type: boolean
      errors:
        items:
          $ref: '#/definitions/utils.ValidatorErrors'
        type: array
      msg:
        example: Validation failed
        type: string
    type: object
  models.Author:
    properties:
      bio:
        type: string
      birth_date:
        type: string
      books:
        items:
          $ref: '#/definitions/models.Book'
        type: array
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.Book:
    properties:
      author:
        $ref: '#/definitions/models.Author'
      author_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      isbn:
        type: string
      publish_year:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  utils.ValidatorErrors:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
host: localhost:3001
info:
  contact: {}
  description: API for managing books, authors, and users with JWT authentication
  title: Fiber Example API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user using username and password
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginAuthDto'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/dto.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Login user
      tags:
      - Auth
  /api/authors:
    get:
      consumes:
      - application/json
      description: Get list of all authors
      produces:
      - application/json
      responses:
        "200":
          description: List of authors
          schema:
            $ref: '#/definitions/dto.GetAuthorsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      security:
      - Bearer: []
      summary: Get all authors
      tags:
      - Authors
    post:
      consumes:
      - application/json
      description: Create a new author with the provided information
      parameters:
      - description: Author information
        in: body
        name: author
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAuthorRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Author created successfully
          schema:
            $ref: '#/definitions/dto.DataResponse-models_Author'
        "400":
          description: Bad request / validation error
          schema:
            $ref: '#/definitions/dto.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      security:
      - Bearer: []
      summary: Create a new author
      tags:
      - Authors
  /api/authors/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an author by ID
      parameters:
      - description: Author ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Author deleted successfully
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "404":
          description: Author not found
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "409":
          description: Cannot delete author with existing books
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      security:
      - Bearer: []
      summary: Delete an author
      tags:
      - Authors
    put:
      consumes:
      - application/json
      description: Update an existing author's information
      parameters:
      - description: Author ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated author information
        in: body
        name: author
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAuthorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Author updated successfully
          schema:
            $ref: '#/definitions/dto.DataResponse-models_Author'
        "400":
          description: Bad request / validation error
          schema:
            $ref: '#/definitions/dto.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "404":
          description: Author not found
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      security:
      - Bearer: []
      summary: Update an author
      tags:
      - Authors
  /api/books:
    get:
      consumes:
      - application/json
      description: Get list of all books
      produces:
      - application/json
      responses:
        "200":
          description: List of books
          schema:
            $ref: '#/definitions/dto.GetBooksResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      security:
      - Bearer: []
      summary: Get all books
      tags:
      - Books
  /api/books/with-authors:
    get:
      consumes:
      - application/json
      description: Get list of all books including their author information
      produces:
      - application/json
      responses:
        "200":
          description: List of books with authors
          schema:
            $ref: '#/definitions/dto.GetBooksWithAuthorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      security:
      - Bearer: []
      summary: Get all books with author details
      tags:
      - Books
  /api/root/revoke/{id}:
    post:
      consumes:
      - application/json
      description: Revoke a user's token version to invalidate all their tokens. Admin
        only.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Token version revoked
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      security:
      - Bearer: []
      summary: Revoke user token version (Admin only)
      tags:
      - Admin
  /api/root/users:
    get:
      consumes:
      - application/json
      description: Get list of all users. Admin only.
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            $ref: '#/definitions/dto.GetUsersResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      security:
      - Bearer: []
      summary: Get all users (Admin only)
      tags:
      - Admin
  /api/users/logout:
    post:
      consumes:
      - application/json
      description: Revoke user's token version to invalidate all tokens
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged out
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      security:
      - Bearer: []
      summary: Logout user
      tags:
      - Users
  /api/users/me:
    get:
      consumes:
      - application/json
      description: Get authenticated user's information
      produces:
      - application/json
      responses:
        "200":
          description: User information
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      security:
      - Bearer: []
      summary: Get current user info
      tags:
      - Users
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"